Make ParquetDataset inherit from base Dataset class
